`yii\jquery\Captcha` extracted
