+++ framework/jquery/ActiveFormAsset.php
@@ -18,7 +18,7 @@ class ActiveFormAsset extends AssetBundle
     /**
      * @inheritdoc
      */
-    public $sourcePath = '@yii/assets';
+    public $sourcePath = '@yii/jquery/assets';
     /**
      * @inheritdoc
      */
